refactor(fs): split fuse dispatcher
